Tasks 1,2,3,4 for Lecture4Homework
