+++ IrConsoleApp.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32210.238
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IrynaZharnasek_SaysHello", "IrConsoleApp\IrynaZharnasek_SaysHello.csproj", "{54E84624-8D9F-4F86-AEB0-0CE6BED9FDE8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IrynaZharnasek", "IrConsoleApp\IrynaZharnasek.csproj", "{54E84624-8D9F-4F86-AEB0-0CE6BED9FDE8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
